requirements.txt
mujoco
google-genai
fastapi
uvicorn[standard]
websockets
python-dotenv
Pillow
numpy
